requirements.txt
aiogram==3.13.1
google-generativeai==0.8.3
python-dotenv==1.0.1
aiosqlite==0.20.0
email-validator==2.2.0
pydantic==2.8.2
Pillow==10.4.0
cryptography==43.0.1
